[REF] spreadsheet: isolate global filter in own plugin
This commit introduce a new ui plugin to manage global filters in pivot.
This is a step to make the pivot ui agnostic of the pivot type.

Task: 3748727
